requirements.txt
streamlit
Pillow
numpy
svgwrite
cairosvg
scikit-image
scikit-learn
